# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Code Review Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repositories:
    get:
      operationId: listRepositories
      tags: [repositories]
      summary: List all repositories
      parameters:
        - name: language
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repository"
  /repositories/{id}:
    get:
      operationId: getRepository
      tags: [repositories]
      summary: Get a repository by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Repository details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all reviews
      parameters:
        - name: repositoryId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a specific review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}/files:
    get:
      operationId: getReviewFiles
      tags: [reviews]
      summary: Get files changed in a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Files changed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewFile"
  /reviews/{id}/comments:
    get:
      operationId: getReviewComments
      tags: [reviews]
      summary: Get AI comments for a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewComment"
...
  /reviews/history:
    get:
      operationId: getReviewHistory
      tags: [reviews]
      summary: Get review history with search
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Review history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewHistoryItem"
components:
  schemas:
...
    Repository:
      type: object
      required: [id, name, fullName, language, stars, openPrs, lastReviewedAt, reviewScore]
      properties:
        id:
          type: integer
        name:
          type: string
        fullName:
          type: string
        language:
          type: string
        stars:
          type: integer
        openPrs:
          type: integer
        lastReviewedAt:
          type: ["string", "null"]
        reviewScore:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        isPrivate:
          type: boolean
...
    Review:
      type: object
      required: [id, pullRequestId, pullRequestNumber, repositoryName, author, qualityScore, securityScore, complexityScore, documentationScore, status, reviewedAt]
      properties:
        id:
          type: integer
        pullRequestId:
          type: integer
        pullRequestNumber:
          type: integer
        repositoryName:
          type: string
        author:
          type: string
        authorAvatar:
          type: ["string", "null"]
        qualityScore:
          type: number
        securityScore:
          type: number
        complexityScore:
          type: number
        documentationScore:
          type: number
        status:
          type: string
        reviewedAt:
          type: string
        aiSummary:
          type: ["string", "null"]
        topIssues:
          type: array
          items:
            type: string
        modelUsed:
          type: ["string", "null"]
        inputTokens:
          type: ["integer", "null"]
        outputTokens:
          type: ["integer", "null"]
        cost:
          type: ["number", "null"]
        agentOutputs:
          type: ["string", "null"]
        promptTemplates:
          type: ["string", "null"]
        latencyMs:
          type: ["integer", "null"]
    ReviewFile:
      type: object
      required: [id, reviewId, filePath, additions, deletions, status]
      properties:
        id:
          type: integer
        reviewId:
          type: integer
        filePath:
          type: string
        additions:
          type: integer
        deletions:
          type: integer
        status:
          type: string
          enum: [added, modified, deleted]
        content:
          type: ["string", "null"]
    ReviewComment:
      type: object
      required: [id, reviewId, filePath, line, type, severity, message]
      properties:
        id:
          type: integer
        reviewId:
          type: integer
        filePath:
          type: string
        line:
          type: integer
        type:
          type: string
          enum: [bug, security, complexity, documentation, style, performance]
        severity:
          type: string
          enum: [critical, high, medium, low]
        message:
          type: string
        suggestion:
          type: ["string", "null"]
        codeSnippet:
          type: ["string", "null"]
...
    ReviewHistoryItem:
      type: object
      required: [id, pullRequestNumber, repositoryName, score, status, reviewedAt]
      properties:
        id:
          type: integer
        pullRequestNumber:
          type: integer
        repositoryName:
          type: string
        score:
          type: number
        status:
          type: string
        reviewedAt:
          type: string
        author:
          type: ["string", "null"]
...