# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Code Review Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  pullRequestId: number;
  pullRequestNumber: number;
  repositoryName: string;
  author: string;
  /** @nullable */
  authorAvatar?: string | null;
  qualityScore: number;
  securityScore: number;
  complexityScore: number;
  documentationScore: number;
  status: string;
  reviewedAt: string;
  /** @nullable */
  aiSummary?: string | null;
  topIssues?: string[];
  /** @nullable */
  modelUsed?: string | null;
  /** @nullable */
  inputTokens?: number | null;
  /** @nullable */
  outputTokens?: number | null;
  /** @nullable */
  cost?: number | null;
  /** @nullable */
  agentOutputs?: string | null;
  /** @nullable */
  promptTemplates?: string | null;
  /** @nullable */
  latencyMs?: number | null;
}